既存storieのディレクトリ変更
