+++ src/stories/server/Genre.stories.tsx
@@ -1,11 +1,11 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { Slider } from "@/app/_components/layouts/Slider";
-import { Jacket } from "@/app/_components/server/Jacket/Jacket";
+import { Genre } from "@/app/_components/server/Genre/Genre";
 
 const meta = {
-	title: "Client/Jacket",
-	component: Jacket,
+	title: "Server/Genre",
+	component: Genre,
 	parameters: {
 		layout: "centered",
 		nextjs: {
@@ -13,64 +13,56 @@ const meta = {
 		},
 	},
 	tags: ["autodocs"],
-} satisfies Meta<typeof Jacket>;
+} satisfies Meta<typeof Genre>;
 
 export default meta;
-type Story = StoryObj<typeof Jacket>;
+type Story = StoryObj<typeof Genre>;
 
-const DUMMY_IMAGES = [
+const DUMMY_GENRES = [
 	{
 		id: 1,
 		src: "/images/dummy-image.png",
-		albumName: "music",
-		artistName: "artist",
+		genreName: "genre",
 	},
 	{
 		id: 2,
 		src: "/images/dummy-image.png",
-		albumName: "music",
-		artistName: "artist",
+		genreName: "genre",
 	},
 	{
 		id: 3,
 		src: "/images/dummy-image.png",
-		albumName: "music",
-		artistName: "artist",
+		genreName: "genre",
 	},
 	{
 		id: 4,
 		src: "/images/dummy-image.png",
-		albumName: "music",
-		artistName: "artist",
+		genreName: "genre",
 	},
 	{
 		id: 5,
 		src: "/images/dummy-image.png",
-		albumName: "music",
-		artistName: "artist",
+		genreName: "genre",
 	},
 	{
 		id: 6,
 		src: "/images/dummy-image.png",
-		albumName: "music",
-		artistName: "artist",
+		genreName: "genre",
 	},
 	{
 		id: 7,
 		src: "/images/dummy-image.png",
-		albumName: "music",
-		artistName: "artist",
+		genreName: "genre",
 	},
 ];
 
 export const Primary: Story = {
 	render: () => (
-		<Jacket
+		<Genre
 			href="/"
 			priority
 			src={"/images/dummy-image.png"}
-			album={{ name: "music", href: "/" }}
-			artist={{ name: "artist", href: "/" }}
+			genreName="genre"
 			width={200}
 			height={200}
 			alt="dummy"
@@ -82,14 +74,13 @@ export const WithSlider: Story = {
 	render: () => (
 		<div style={{ width: "1000px" }}>
 			<Slider>
-				{DUMMY_IMAGES.map((image) => (
-					<Jacket
-						key={image.id}
+				{DUMMY_GENRES.map((genre) => (
+					<Genre
+						key={genre.id}
 						href="/"
 						priority
-						src={image.src}
-						album={{ name: image.albumName, href: "/" }}
-						artist={{ name: image.artistName, href: "/" }}
+						src={genre.src}
+						genreName={genre.genreName}
 						width={200}
 						height={200}
 						alt="dummy"
